Clean up test_makepreservation.py
Split tests into one document per function.
Edited comments.
