+++ tests/expected_xml/aip-id_preservation.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<preservation xmlns:dc="http://purl.org/dc/terms/"
-              xmlns:premis="http://www.loc.gov/premis/v3">
+<preservation xmlns:premis="http://www.loc.gov/premis/v3"
+              xmlns:dc="http://purl.org/dc/terms/">
    <dc:title>title</dc:title>
    <dc:rights>http://rightsstatements.org/vocab/InC/1.0/</dc:rights>
    <aip>
